feat: wire scheduled Causal Geometry Phase B producer + shared postgres-disk store path
The HITL proposal queue Phase B shipped in PR #1087 was real but permanently
empty -- nothing ever called propose_field_topology_patches()/store.propose()
in production. This wires the missing producer and fixes the learned-weights
store's persistence path per operator direction (shared disk alongside
Postgres, not /mnt/telemetry).

- orion/substrate/causal_geometry_engine.py (new): Phase A's core measurement
  logic (fetch_channels, bucketize, lagged correlation, Bonferroni-corrected
  surrogate significance, divergence) extracted from
  scripts/causal_geometry_report.py into an importable module, so the new
  producer can call the exact same code the CLI report uses instead of
  duplicating it. scripts/causal_geometry_report.py now re-exports every name
  it used to define, so its own test suite (tests/test_causal_geometry_report.py)
  is unaffected.
- orion/substrate/causal_geometry_producer.py (new): run_causal_geometry_production_cycle()
  -- measures a snapshot, proposes field_topology_weight_patch candidates, and
  enqueues only those without an already-pending duplicate for the same edge.
  Never raises (degrades to an {"ok": False, ...} summary on any failure,
  matching diffusion.py's _load_learned_overlay() degradation pattern). Never
  imports orion.substrate.mutation_apply (AST-guarded by its own test, same
  hard constraint as field_topology_plasticity.py).
- services/orion-field-digester/app/worker.py: new _causal_geometry_producer_loop
  (same asyncio.to_thread + cancellable-sleep shape as the existing prune/health
  loops), off by default via FIELD_PLASTICITY_PRODUCER_ENABLED, calling the
  producer on a configurable interval (default 24h, matching the design spec's
  "deterministic nightly job" framing) with the digester's own already-loaded
  lattice edges and Postgres URI.
- services/orion-field-digester/app/digestion/diffusion.py: renamed
  _get_learned_store -> get_learned_store (public) so worker.py's producer and
  diffusion.py's read path share one store instance/sqlite connection per
  process instead of each building their own.
- FIELD_PLASTICITY_SQL_DB_PATH now defaults to a real path,
  ${STORAGE_ROOT:-/mnt/postgres}/field_topology/learned_weights.sqlite3 (a
  sibling directory to postgres_data, never inside it), instead of empty --
  per operator direction to live on the same durable disk as Postgres rather
  than /mnt/telemetry. Both services' docker-compose.yml now mount
  ${STORAGE_ROOT:-/mnt/postgres}/field_topology into the same container path
  so both containers see the same sqlite file.
- services/orion-field-digester/requirements.txt: added numpy==1.26.4
  (matches the version pinned by most other services in this repo) --
  the engine module now imported into this service needs it.

Tests: orion/substrate/tests (332, +6 new), orion/schemas/tests, and
tests/test_causal_geometry_report.py all pass unchanged. orion-field-digester's
own suite: 63 passed (+3 new worker-level tests covering disabled/enabled/
failure-degradation). orion-hub's causal-geometry-scoped tests: 22 passed,
unaffected. docker compose config validated for both services (nested
${STORAGE_ROOT:-/mnt/postgres} expansion inside an .env-sourced value
confirmed empirically to resolve correctly, matching the existing
orion-sql-db/.env_example VOLUMES_ORION_POSTGRES_PATH precedent).
