+++ services/orion-field-digester/.env_example
@@ -70,6 +70,23 @@ FIELD_PLASTICITY_DECAY_HALF_LIFE_HOURS=168
 # adopting a proposal via the orion-hub tab will NOT be visible here, since hub and
 # field-digester are separate containers with separate in-memory stores. Must be set
 # to the SAME path as orion-hub's FIELD_PLASTICITY_SQL_DB_PATH, on a volume mounted
-# into both containers (e.g. under the shared ${TELEMETRY_ROOT:-/mnt/telemetry} mount
-# this service already uses), for cross-process adoption to actually reach diffusion.
-FIELD_PLASTICITY_SQL_DB_PATH=
+# into both containers, for cross-process adoption to actually reach diffusion.
+# Lives on the same durable disk as Postgres (${STORAGE_ROOT:-/mnt/postgres}, see root
+# .env_example), in a sibling directory -- never inside postgres_data itself -- so this
+# small, low-write-volume sqlite file doesn't depend on ephemeral container storage or
+# an unrelated volume's retention/rotation policy.
+FIELD_PLASTICITY_SQL_DB_PATH=${STORAGE_ROOT:-/mnt/postgres}/field_topology/learned_weights.sqlite3
+# Scheduled Phase A -> Phase B producer (orion/substrate/causal_geometry_producer.py):
+# periodically measures observed-vs-designed field topology and enqueues new HITL
+# proposals. Independent of FIELD_PLASTICITY_ENABLED above (that one gates *consuming*
+# the learned overlay in diffusion; this one gates *producing* proposals into the
+# queue). Off by default -- an operator must opt in explicitly.
+FIELD_PLASTICITY_PRODUCER_ENABLED=false
+# Cadence in hours between production cycles. 24h matches the design spec's
+# "deterministic nightly job" framing -- this is a full Postgres pull plus
+# O(channels^2) lagged-correlation + surrogate-significance computation, not
+# something to run on this service's own ~2s tick cadence.
+FIELD_PLASTICITY_PRODUCER_INTERVAL_HOURS=24
+# Lookback window in hours for each production cycle. Matches
+# orion/substrate/causal_geometry_engine.py's DEFAULT_WINDOW_HOURS (168h / 7 days).
+FIELD_PLASTICITY_PRODUCER_WINDOW_HOURS=168
