+++ services/orion-field-digester/requirements.txt
@@ -4,5 +4,6 @@ pydantic==2.10.3
 pydantic-settings==2.6.1
 sqlalchemy==2.0.36
 psycopg2-binary==2.9.10
+numpy==1.26.4
 PyYAML==6.0.2
 requests==2.31.0
