+++ services/orion-hub/.env_example
@@ -437,8 +437,10 @@ SUBSTRATE_MUTATION_OPERATOR_TOKEN=
 # in-memory only, this process only -- adopting a proposal here will NOT reach
 # orion-field-digester's diffusion tick. Must be set to the SAME path as
 # orion-field-digester's FIELD_PLASTICITY_SQL_DB_PATH, on a volume mounted into both
-# containers, for an adoption to actually take effect.
-FIELD_PLASTICITY_SQL_DB_PATH=
+# containers, for an adoption to actually take effect. Lives on the same durable disk
+# as Postgres (${STORAGE_ROOT:-/mnt/postgres}, see root .env_example), in a sibling
+# directory -- never inside postgres_data itself.
+FIELD_PLASTICITY_SQL_DB_PATH=${STORAGE_ROOT:-/mnt/postgres}/field_topology/learned_weights.sqlite3
 
 # --- Substrate semantic graph (Fuseki / SPARQL or legacy GraphDB) ---
 # RDF Store V1: Hub docker-compose defaults SUBSTRATE_STORE_BACKEND=sparql with Athena Fuseki URLs; override for host mode (127.0.0.1).
